Reorganize styles alongside components
