+++ src/components/Base/Header.css
@@ -1,0 +1,8 @@
+.Header {
+  min-height: 80px;
+  display: flex;
+  flex-direction: row;
+  align-items: center;
+  justify-content: space-between;
+  padding: 10px 30px;
+}
