+++ src/components/Base/Player.css
@@ -1,0 +1,55 @@
+.Player {
+  width: 85px;
+  height: 142px;
+  background: #212121;
+  border: 1px solid rgb(151, 151, 151);
+  border-radius: 5px;
+  font-size: 1rem;
+  display: inline-flex;
+  justify-content: center;
+  align-items: center;
+  flex-direction: column;
+  position: relative;
+  margin: 0.3rem;
+  color: #f5f5f5;
+  padding: 0;
+  border: 3px solid transparent;
+  cursor: pointer;
+}
+.Player:disabled {
+  cursor: default;
+}
+
+.avatar {
+  max-width: 100%;
+  width: 2.8rem;
+  margin-bottom: 0.2rem;
+}
+
+.Player.selected {
+  outline: 0;
+  border: 3px solid #ffbb01;
+}
+
+.Player:focus {
+  outline: 0;
+}
+
+.Player .hammer-icon {
+  position: absolute;
+  top: 3px;
+  right: 3px;
+  width: 18px;
+}
+
+.Player .king-icon {
+  position: absolute;
+  top: 8px;
+}
+
+.Player .vote-icon {
+  position: absolute;
+  bottom: 7px;
+  width: 16px;
+  height: 16px;
+}
