+++ src/components/GameBoard/ActionArea/ActionArea.css
@@ -1,0 +1,3 @@
+.ActionArea {
+  margin-bottom: 5rem;
+}
