+++ src/components/GameBoard/ActionArea/MissionVote/MissionVote.css
@@ -1,0 +1,3 @@
+.MissionVote {
+  position: relative;
+}
