+++ src/components/GameBoard/ActionArea/MissionVote/MissionVote.jsx
@@ -1,5 +1,6 @@
 import React from 'react';
 import Card from '../../../Base/Card';
+import './MissionVote.css';
 import {dispatchSubmitMissionVote} from '../../../../ApiUtils';
 
 class MissionVote extends React.Component {
