+++ src/components/GameBoard/ActionArea/TeamProposal/VoteResultView.css
@@ -1,0 +1,3 @@
+.Player-list {
+  margin: 20px 0;
+}
