+++ src/components/GameBoard/Board.css
@@ -1,0 +1,15 @@
+/* Portrait and Landscape */
+@media (max-width: 767px) {
+    .Board:before {
+        content: '';
+        pointer-events: none;
+        width: 100%;
+        height: 100%;
+        position: absolute;
+        background: transparent;
+        left: 0;
+        top: 0;
+        z-index: 10;
+        box-shadow: inset 0px -20px 25px 20px black;
+    }
+}
