+++ src/components/PreGame/CreateRoom.css
@@ -1,0 +1,18 @@
+.Create-room {
+  display: flex;
+  flex-direction: column;
+  max-width: 300px;
+  margin: auto;
+}
+
+.MuiInput-root.Form-select {
+  margin-top: 20px;
+}
+
+.MuiFormLabel-root.Form-label {
+  margin-top: 30px;
+}
+
+.Form-submit {
+  margin: 30px 0;
+}
