+++ src/components/PreGame/CreateRoom.jsx
@@ -11,7 +11,7 @@ import TextField from "@material-ui/core/TextField";
 import Button from '../Base/Button';
 import { createRoom } from '../../ApiUtils';
 
-export default class CreateForm extends React.Component {
+export default class CreateRoom extends React.Component {
   constructor(props) {
     super(props);
