+++ src/components/PreGame/JoinRoom.css
@@ -1,0 +1,7 @@
+.Join-room {
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  flex-direction: column;
+  margin-top: 20px;
+}
