+++ src/components/styles.css
@@ -1,119 +1,4 @@
-/* TODO reorganize styles into component folders */
-
-.Header {
-  min-height: 80px;
-  display: flex;
-  flex-direction: row;
-  align-items: center;
-  justify-content: space-between;
-  padding: 10px 30px;
-}
-
-.Button {
-  min-width: 100px;
-  background-color: #00d673;
-  padding: 10px 15px;
-  border: none;
-  border-radius: 15px;
-  font-weight: 500;
-  font-size: .8rem;
-  cursor: pointer;
-  margin: 10px 0;
-}
-
-.Button:disabled {
-  cursor: default;
-}
-
-.Text-P {
-  font-size: 1.25rem;
-  font-weight: 500;
-  margin: 0;
-  padding: 0;
-}
-
-.Text-Heading {
-  font-size: 2rem;
-  font-weight: 600;
-  margin: 10px 0;
-  padding: 0;
-}
-
-.Text-Sub {
-  font-size: 1rem;
-  font-weight: 400;
-  margin: 0;
-  padding: 0;
-}
-
-.Text-Highlight {
-  font-size: 1.25rem;
-  font-weight: 550;
-  display: block;
-  text-decoration: underline;
-  margin-bottom: 5px;
-  padding: 0;
-  color: #ccc;
-}
-
-
-.Text-Bold {
-  font-size: 1.25rem;
-  font-weight: 550;
-  margin-bottom: 5px;
-  display: block;
-  padding: 0;
-  color: #ccc;
-}
-
-.Text-Link {
-  font-size: 1rem;
-  font-weight: 400;
-  margin: 0;
-  padding: 0;
-  text-decoration: underline;
-  color: #ccc;
-}
-
-.Text-Link:hover {
-  text-decoration: underline;
-  color: white;
-}
-
-.Mission-wrapper {
-  width: 105px;
-  height: 110px;
-  border-radius: 5px;
-  margin: 0 20px;
-  padding: 10px;
-  display: inline-block;
-  position: relative;
-}
-
-.Mission-wrapper.active {
-  background-color: #212121;
-}
-
-.Mission {
-  border-radius: 50%;
-  background-color: #555555;
-  font-size: 3rem;
-  width: 102px;
-  height: 102px;
-  margin: 0;
-  display: flex;
-  align-items: center;
-  justify-content: center;
-}
-
-.Mission.success {
-  background-color: #015aff;
-}
-
-.Mission.fail {
-  background-color: #d10146;
-}
-
+/* Shared styles only */
 .red {
   color: #d10146;
 }
@@ -122,208 +7,21 @@
   color: #015aff
 }
 
-.Mission-tracker {
-  border-radius: 50%;
-  background-color: #555555;
-  width: 3rem;
-  height: 3rem;
-  position: absolute;
-  left: 0;
-  bottom: 5px;
-  font-size: .9rem;
-  display: flex;
-  align-items: center;
-  justify-content: center;
+.PlayerGroup {
+  margin-bottom: 1rem;
 }
 
-.Mission-tracker {
-  margin-left: -0.7rem;
-  margin-bottom: -0.7rem;
-  box-shadow: 0 0 0 3px rgba(0,0,0,0.1);
+.PlayerGroup--KingView {
+  margin-bottom: 1rem;
 }
 
-.Mission-tracker:after {
-  content: ' attempts';
-  display: inline-block;
-  position: absolute;
-  bottom: -0.2rem;
+.PlayerGroup--PlayerView .Player {
+  pointer-events: none;
 }
 
-
-.Mission-double-fail {
-  font-size: .8rem;
-}
-
-.Mission-tracker.hammer-vote {
-  background-color: #d10146;
-}
-
-.Card-list,
-.MissionVote {
-  position: relative;
-}
-
-.Card {
-  width: 135px;
-  height: 225px;
-  background: rgb(85, 85, 85);
-  border-radius: 5px;
-  color: #212121;
-  margin: 7px;
-  display: inline-flex;
-  flex-direction: column;
-  justify-content: center;
-  align-items: center;
-  border: 3px solid transparent;
-}
-
-.CardWrapper {
-  display: inline-block;
-}
-
-.CardWrapper input {
-  opacity: 0;
-  position: absolute;
-  top: 0;
-  width: 135px;
-  height: 225px;
-}
-
-.CardWrapper input:checked + .Card {
-  border: 3px solid #ffbb01;
-}
-.CardWrapper input.success:checked + .Card svg {
-  z-index: 2;
-}
-
-.CardWrapper input:checked + .fail:after,
-.CardWrapper input:checked + .success:after {
-  position: absolute;
-  top: 0;
-  left: 0;
-  content: '';
-  background-color: black;
-  opacity: 0.6;
-  z-index: 1;
-  width: 100%;
-  height: 100%;
-}
-
-.Player {
-  width: 85px;
-  height: 142px;
-  background: #212121;
-  border: 1px solid rgb(151, 151, 151);
-  border-radius: 5px;
-  font-size: 1rem;
-  display: inline-flex;
-  justify-content: center;
-  align-items: center;
-  flex-direction: column;
-  position: relative;
-  margin: 0.3rem;
-  color: #f5f5f5;
-  padding: 0;
-  border: 3px solid transparent;
-  cursor: pointer;
-}
-.Player:disabled {
-  cursor: default;
-}
-
-.avatar {
-  max-width: 100%;
-  width: 2.8rem;
-  margin-bottom: 0.2rem;
-}
-
-.Player.selected {
-  outline: 0;
-  border: 3px solid #ffbb01;
-}
-
-.Player:focus {
-  outline: 0;
-}
-
-.Player .hammer-icon {
-  position: absolute;
-  top: 3px;
-  right: 3px;
-  width: 18px;
-}
-
-.Player .king-icon {
-  position: absolute;
-  top: 8px;
-}
-
-.Player .vote-icon {
-  position: absolute;
-  bottom: 7px;
-  width: 16px;
-  height: 16px;
-}
-
-.Join-room {
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  flex-direction: column;
-  margin-top: 20px;
-}
-
-.MuiDrawer-paperAnchorRight {
-  width: 380px;
-  border-radius: 5px;
-  background-color: #eee;
-  padding: 20px;
-}
-
-.Drawer-close {
-  cursor: pointer;
-  color: red;
-}
-
-.Info-icon {
-  cursor: pointer;
-}
-
-.MuiTabs-root {
-  background-color: #212121;
-}
-
-.Rules-role {
-  margin-bottom: 20px;
-  padding-bottom: 15px;
-  border-bottom: 1px solid #555555;
-}
-
-.Create-room {
-  display: flex;
-  flex-direction: column;
-  max-width: 300px;
-  margin: auto;
-}
-
-.MuiInput-root.Form-select {
-  margin-top: 20px;
-}
-
-.MuiFormLabel-root.Form-label {
-  margin-top: 30px;
-}
-
-.Form-submit {
-  margin: 30px 0;
-}
-
-.Player-list {
-  margin: 20px 0;
-}
-
-@media screen and (max-width: 600px) {
-  .Button {
-    width: 75%;
-  }
+/* TODO put this style on ActionArea instead? or move Mission-result next to its component */
+.LayoutGroup--KingProposalView,
+.Vote-Result,
+.Mission-Result {
+  margin-bottom: 5rem;
 }
